requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyMuPDF==1.23.8
Pillow==10.1.0
python-magic-bin==0.4.14
